requirements.txt
openai
crewai
langgraph
pyautogen
mcp
google-genai
python-dotenv
jupyter
gradio
pypdf
rich
requests
pydantic